feat: add vscode config, utils from .js to .ts
